Add files via upload
Email spam classification is a machine learning project which aims to classify the mails into spam or ham based on any classifier. The classifier used in this project is Multinomial Naive Bayes classifier and produced an accuracy of 97% in its classification. It also correctly predicted the category of 'spam' or 'ham' when given the new data as input.