Add files via upload
Social Network Analysis is a part of machine learning and NLP techniques used to identify the network connections within the members of a group. The project is implemented on Zachary Karate Club Dataset to identify the connections and visualize it in the form of graph. Also, various measures of SNA such as degree centrality, closeness centrality and betweeness centrality were measured in this project.